Make benches more modular
