+++ benches/ring.rs
